cleanup, moved metrics to types
