+++ block/produce.go
@@ -157,9 +157,9 @@ func (m *Manager) produceBlock(ctx context.Context, allowEmpty bool) error {
 	}
 
 	m.logger.Info("block created", "height", newHeight, "num_tx", len(block.Data.Txs))
-	rollappBlockSizeBytesGauge.Set(float64(len(block.Data.Txs)))
-	rollappBlockSizeTxsGauge.Set(float64(len(block.Data.Txs)))
-	rollappHeightGauge.Set(float64(newHeight))
+	types.RollappBlockSizeBytesGauge.Set(float64(len(block.Data.Txs)))
+	types.RollappBlockSizeTxsGauge.Set(float64(len(block.Data.Txs)))
+	types.RollappHeightGauge.Set(float64(newHeight))
 	return nil
 }
