+++ types/metrics.go
@@ -1,23 +1,23 @@
-package block
+package types
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 )
 
-var rollappHeightGauge = promauto.NewGauge(prometheus.GaugeOpts{
+var RollappHeightGauge = promauto.NewGauge(prometheus.GaugeOpts{
 	Name: "rollapp_height",
 	Help: "The height of the local rollapp",
 })
-var rollappHubHeightGauge = promauto.NewGauge(prometheus.GaugeOpts{
+var RollappHubHeightGauge = promauto.NewGauge(prometheus.GaugeOpts{
 	Name: "rollapp_hub_height",
 	Help: "The latest height of the Rollapp that has been synced to the hub.",
 })
-var rollappBlockSizeBytesGauge = promauto.NewGauge(prometheus.GaugeOpts{
+var RollappBlockSizeBytesGauge = promauto.NewGauge(prometheus.GaugeOpts{
 	Name: "rollapp_block_size_bytes",
 	Help: "Rollapp ",
 })
-var rollappBlockSizeTxsGauge = promauto.NewGauge(prometheus.GaugeOpts{
+var RollappBlockSizeTxsGauge = promauto.NewGauge(prometheus.GaugeOpts{
 	Name: "rollapp_block_size_txs",
 	Help: "Rollapp ",
 })
